added

+++ 1.scala
@@ -10,9 +10,9 @@ object demo{
 
     def main(args:Array[String]):Unit={
         println(k+12*m)
-        println(m/j)
+        println(m.toFloat/j)
         println(n%j)
-        println(m/j*j)
+        println(m.toFloat/(j*j))
         println(f+10*5+g)
         i=i+1
         println(i*n)
